♻️ refactor : `@Repository` 어노테이션 제거
