+++ src/main/java/com/spring/mummus/member/repository/MemberRepository.java
@@ -1,13 +1,9 @@
-package com.spring.mummus.repository;
+package com.spring.mummus.member.repository;
 
 
-import com.spring.mummus.domain.entity.Member;
+import com.spring.mummus.member.domain.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
-import java.util.Optional;
-
-@Repository
 public interface MemberRepository extends JpaRepository<Member, Long> {
     Boolean existsByEmail(String email);
     Boolean existsByPhoneNumber(String phoneNumber);
