+++ src/main/java/com/spring/mummus/pet/repository/PetRepository.java
@@ -1,10 +1,8 @@
-package com.spring.mummus.repository;
+package com.spring.mummus.pet.repository;
 
-import com.spring.mummus.domain.entity.Pet;
+import com.spring.mummus.pet.domain.entity.Pet;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
-@Repository
 public interface PetRepository extends JpaRepository<Pet, Long> {
     Boolean existsByNameAndMemberId(String name, Long memberId);
 }
